KNN Graphs

Approximate KNN Graphs

Functions

NearestNeighborDescent.KNNGraphs.knn_diameter — Function

knn_diameter(g::ApproximateKNNGraph{V}, v::V)

Compute the diameter of the ball centered on v that covers all of vs approximate k-nearest neighbors.

knn_diameter(graph, v) -> diameter

Return the diameter of the set of KNNs of vertex v.

NearestNeighborDescent.KNNGraphs.knn_matrices — Function

knn_matrices(graph) -> indices, distances

Return the indices and distances of the approximate KNNs as dense matrices where indices[j, i] and distances[j, i] are the index of and distance to node i's jth nearest neighbor, respectively.

Missing docstring.

Missing docstring for flag. Check Documenter's build log for details.

Missing docstring.

Missing docstring for weight. Check Documenter's build log for details.

NearestNeighborDescent.KNNGraphs.edge_indices — Function

edge_indices(graph) -> CartesianIndices

Return the indices of the KNNs for each node v as tuples (v, i). To be used with node_edge(graph, v, i).

NearestNeighborDescent.KNNGraphs.node_edge — Function

node_edge(graph, v, i) -> edge

Return the ith outgoing edge from node v. No ordering of the edges is guaranteed; in particular, node_edge(graph, v, 1) is not guaranteed to be the edge to v's nearest neighbor.

NearestNeighborDescent.KNNGraphs.node_edges — Function

node_edges(graph::LockHeapKNNGraph, i) -> edges

Return all the outgoing edges from node i in an arbitrary order. Thread-safe.

node_edges(graph, v) -> edges

Return all the outgoing edges from node v in an arbitrary order.

NearestNeighborDescent.KNNGraphs.update_flag! — Function

update_flag!(g::LockHeapKNNGraph, i, j, flag)

Update the flag of the edge at the given indices. Since the flags don't influence the edge ordering, this can't invalidate the heap invariant. Uses locks to ensure thread safety.

update_flag!(graph, v, i, new_flag)

Update the flag of node vs ith outgoing edge. Returns the new edge. Note that since the flags don't influence the edge ordering, this can't invalidate the heap invariant.

Types

NearestNeighborDescent.KNNGraphs.ApproximateKNNGraph — Type

ApproximateKNNGraph{V, K} subtypes are weighted, directed graphs where each vertex has exactly k forward edges.

NearestNeighborDescent.KNNGraphs.HeapKNNGraph — Type

HeapKNNGraph{V, K, U}

A weighted, directed graph representing an approximate k-nearest neighbors graph using binary max heaps to store each vertex's forward edges, allowing for efficient updates of the candidate neighbors.

NearestNeighborDescent.KNNGraphs.LockHeapKNNGraph — Type

LockHeapKNNGraph - uses locks to synchronize the heaps that store the underlying graph edge data. The heaps themselves are not thread-safe.

NearestNeighborDescent.KNNGraphs.HeapKNNGraphEdge — Type

HeapKNNGraphEdge{T, U}

A weighted graph edge along with a flag